requirements.txt
amqp==5.3.1
asgiref==3.9.1
billiard==4.2.1
celery==5.5.3
certifi==2025.8.3
cffi==1.17.1
channels==4.3.1
channels_redis==4.3.0
charset-normalizer==3.4.3
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cryptography==45.0.7
dj-rest-auth==7.0.1
Django==5.2.5
django-allauth==65.11.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
idna==3.10
kombu==5.5.4
msgpack==1.1.1
oauthlib==3.3.1
packaging==25.0
pillow==11.3.0
prompt_toolkit==3.0.52
pycparser==2.22
PyJWT==2.10.1
python-dateutil==2.9.0.post0
redis==6.4.0
requests==2.32.5
rest-framework-simplejwt==0.0.2
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.13
